Relocate deleteFeedbackSessionsDeadlinesForUser to deadline extensions logic
